+++ packages/generator/src/generators/extractCompact.ts
@@ -1,0 +1,43 @@
+/**
+ * @copyright   2017-2018, Miles Johnson
+ * @license     https://opensource.org/licenses/MIT
+ */
+
+import { TEXT, CompactEmoji, Emoji } from 'emojibase';
+
+export default function extractCompact(data: Emoji[]): CompactEmoji[] {
+  return data.map(full => {
+    const {
+      hexcode,
+      shortcodes,
+      emoji,
+      emoticon,
+      order,
+      group,
+      annotation,
+      tags,
+      text,
+      type,
+      skins,
+    } = full;
+    const compact: CompactEmoji = {
+      annotation,
+      group,
+      hexcode,
+      order,
+      shortcodes,
+      tags,
+      unicode: text && type === TEXT ? text : emoji,
+    };
+
+    if (emoticon) {
+      compact.emoticon = emoticon;
+    }
+
+    if (skins) {
+      compact.skins = extractCompact(skins);
+    }
+
+    return compact;
+  });
+}